Implement the basic AnswerListEntry Component
This component will dynamically generate a single answer entry based on
the answer object passed in.

There will be a button allowing the user to increase the helpfulness
rating of this answer by one point but can only be clicked once.

There will be a button that can be clicked once for the user to report
the answer.