Implment Load More Answers button.
Allows the user to reveal more answers to the page via the LOAD MORE
ANSWERS BUTTON.  The number of answers that will be displayed is tracked
via a useState hook.  Clicking the LOAD MORE ANSWERS button will add 2
to this state.  A useEffect hook tracking this state will rerender the
answer list based on this change.