Create AddAnswerModal and AddQuesitonModal components
Both component will be mounted via the Modal Template and thus inherit
those Modal stylings.

Both essentially cretae a form to collect information from the user that
will be submitted to the API to be stored.

AddAnswerModal allows the user to add their own answer to a particular
question for the current product.  It collects the answer body, the
users nickname, the users email, and any photos the user would like to
submit.  Upon clicking the submit button this modal will gather the info
into a post request to the API and will close the modal.

AddQuesitonModal allows the user to add their own question to the list
of questions for the current product.  It collects the question body,
the users nickname, and the users e-mail.  Upon clicking submit this
modal will gather the info into a post request for the API and will
close the modal.