Incorporate searchbar and remove filtering
Searchbar is now implemented directly within this component rather than
built as a separate component that is imported.  This allows for
effciency in processing the search input and passing of the search term
to the question list.  Prviously there was an attempt to filter the list
of questions before passing down to the question list.  This caused
issues wherre the question list would not re-render when a newly
filtered list was created.  Flow has been simplified so all filtering
and sorting happens withi the question list.  This component will track
the searchbar input value and only pass a new one down to the question
list when there is an update relevant to filtering.