Refactor list filtering and sorting based on search term.
Debugged and streamlined process for live filtering of the question list
based on the serach term passed down from the parent component.

Upon a change in the searchInput or the numToDisplay state a useEffect hook will - copy the full
list of questions for the product - filter it based on the search term
if the search term in 3 chars or longer - sort it based on helpfullness
and display the number of questions specified by the current numToDisplay state.

It will also track the currentyl available questions based on the
current search filter so that the component will know whether or not to
make the LOAD MORE QUESTIONS button available.

This commit also eliminates the nesting of conditional return statement
for the render and instead uses a sequential chain of if statements.