Implement helpfullness and report buttons
Upon click this button will increase the questions helpfullness rating
by 1.  This will be reflected immediately upon cicking.  This is only a
superficial display of the helpfullness rating.  Refreshing the page
will get the official updated rating from the server.  Clicking will
also invoke a to be written function that will make a post request to
the API to record the increase in the helpfulness rating on the
database.

The report button will function in the same way.

Both items will have their display text changed to indicate to the user
that the button has been clicked.  These buttons can only be clicked
with effect once.