Add reporting option and API calls for reports/helpful
Adds a button for user to report a question.

Implements the api calls that result from teh already implemented
buttons for reporting a question or marking a question as helpful.
These function will take in the id of the question clicked and create an
axios.put request to the Questions and Answers API.