Update components with containers for API calls.
All components making API calls now have a redux Container.  This will
allow them to update the store with a new list of questions when
required due to a user interaction that mutates or adds data to the
databse.  This is done via a dispatcher that is provided by the
container and then passed into the functions making the API calls where
necessary.