Create InteractionTrakcer component
Creates a compnent that will track user interactions on the page.
Interaction tracker will wrap around the item to be interacted with.
The item to be interacted with is passed into the Interactions tracker
as its render prop.  BY wrapping the item to be interacted with in a
function you can provide the item with the postInteraction() method from
the Interaction tracker.  Using this method will set up a post request
to the interaction tracking api.  It will expect an element and widget
prop to also be set for the instance of the Interaction tracker.  The
widget prop will indicate which widget of the site is being interacted
with.  The element prop will indicate which specific element of the
widget is being interacted with.